#233 동작하지 않는 테스트용 메뉴를 삭제한다
