+++ src/components/templates/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {SmileOutlined} from '@ant-design/icons';
+import {Result} from "antd";
+
+export const Home = () => {
+  return (
+    <Result
+      icon={<SmileOutlined/>}
+      title="Welcome To Better Admin DEMO"
+    />
+  )
+};
+
+
