+++ src/config/navigation.js
@@ -1,7 +1,4 @@
-import {BiBeer, BiBluetooth} from "react-icons/bi";
-import {SiAccusoft, SiAmazon, SiAmazondynamodb, SiApachenetbeanside} from "react-icons/si";
-import {GiAbstract023} from "react-icons/gi";
-import {AiFillSetting, AiOutlineControl} from "react-icons/ai";
+import {AiFillSetting, AiOutlineControl, AiOutlineHome} from "react-icons/ai";
 import {FiUsers} from "react-icons/fi";
 import {TbWebhook} from "react-icons/tb";
 import {MdLogin} from "react-icons/md";
@@ -9,41 +6,9 @@ import {MdLogin} from "react-icons/md";
 export const Navigation = {
   items: [
     {
-      title: "GNB1-22",
-      icon: <BiBluetooth/>,
-      items: [
-        {
-          title: "SNB1",
-          icon: <BiBeer/>,
-          items: [
-            {
-              title: "Sub1",
-              icon: <SiAccusoft/>,
-              link: "/sub1"
-            },
-            {
-              title: "Sub2",
-              icon: <SiAmazon/>,
-              link: "/sub2"
-            }
-          ]
-        },
-        {
-          title: "SNB2",
-          icon: <SiAmazondynamodb/>,
-          disabled: true,
-          link: "/snb2"
-        },
-        {
-          title: "Sample",
-          icon: <SiApachenetbeanside/>,
-          link: "/sample-list"
-        }
-      ]
-    }, {
-      title: "GNB2",
-      icon: <GiAbstract023/>,
-      link: "/gnb2"
+      title: "홈",
+      icon: <AiOutlineHome/>,
+      link: "/home"
     }, {
       title: "접근 제어",
       accessPermissions: ["MANAGE_ACCESS_CONTROL"],
