--- src/pages/Snb2.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Snb2 = () => (
-  <>
-    <div>
-      <h1>Snb2 Page</h1>
-    </div>
-  </>
-);
-
-export default Snb2;
