--- src/pages/samples/SampleDetails.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-function SampleDetails() {
-  return <div>
-    <h1>SampleDetails</h1>
-  </div>;
-}
-
-export default SampleDetails;
